Added .dockerignore file
Using different format for the dockerignore file. Instead of
adding all the files and folders to not add to Docker build context,
exclude all files from build context and then add exceptions for the files
you want in the image. This method seems to be safer as less chance of wrong
files getting placed into image.